Add bench.hpp

+++ bench.hpp
@@ -9,10 +9,10 @@
 namespace avs {
 
 void run_ip_1_x_N(dnnl::engine &engine, dnnl::stream &stream) {
-    int32_t mat_a_size = 1;
-    int32_t mat_a_dim = CONST_A;
-    int32_t mat_b_size = CONST_A;
-    int32_t mat_b_dim = CONST_A;
+    int64_t mat_a_size = 1;
+    int64_t mat_a_dim = CONST_A;
+    int64_t mat_b_size = CONST_A;
+    int64_t mat_b_dim = CONST_A;
 
     std::vector<float> mat_a(mat_a_size * mat_a_dim);
     std::vector<float> mat_b(mat_b_size * mat_b_dim);
@@ -55,10 +55,10 @@ void run_ip_1_x_N(dnnl::engine &engine, dnnl::stream &stream) {
 }
 
 void run_ip_N_x_N(dnnl::engine &engine, dnnl::stream &stream) {
-    int32_t mat_a_size = CONST_A;
-    int32_t mat_a_dim = CONST_A;
-    int32_t mat_b_size = CONST_A;
-    int32_t mat_b_dim = CONST_A;
+    int64_t mat_a_size = CONST_A;
+    int64_t mat_a_dim = CONST_A;
+    int64_t mat_b_size = CONST_A;
+    int64_t mat_b_dim = CONST_A;
 
     std::vector<float> mat_a(mat_a_size * mat_a_dim);
     std::vector<float> mat_b(mat_b_size * mat_b_dim);
